+++ core-libs/setup/ssr/logger/index.ts
@@ -4,8 +4,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './loggers/express-server-logger';
-export * from './loggers/legacy-server-logger';
-export * from './services/express-logger.service';
-export * from './services/prerendering-logger.service';
-export * from './services/server-logger-service-factory';
+export * from './loggers';
+export * from './services';
