+++ core-libs/setup/ssr/logger/loggers/index.ts
@@ -1,8 +1,3 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
+export * from './default-express-server-logger';
 export * from './express-server-logger';
 export * from './legacy-server-logger';
